change triplist component fetching and rendering data
